-API's(Create Beneficiary,Delete,Update) with hard coded value implemented

+++ app/src/main/java/org/jethro/mobile/models/beneficiary/BeneficiaryUpdatePayload.kt
@@ -10,4 +10,4 @@ data class BeneficiaryUpdatePayload @JvmOverloads constructor(
 
     @SerializedName("name") var name: String? = null,
 
-    @SerializedName("transferLimit") var transferLimit: Float = 0f)
+    @SerializedName("transferLimit") var transferLimit: Int = 0)
